+++ CHANGELOG.md
@@ -12,6 +12,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - **accessibility**: 💡 All components now inherit aria-attributes into their shadow DOM and exposes a `setAriaAttribute` method. See new documentation "Accessibility" on Storybook for more info. 
+- **select**: **add server-mode for select autocomplete** (read documentation)
 - **select-option**: add a documentation about handling large text in options
 - **grid**:
   - add support for row styling with css parts